% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treescanr-package.R
\docType{package}
\name{treescanr-package}
\alias{treescanr}
\alias{treescanr-package}
\title{treescanr: Tree-Based Scan Statistics for Syndromic Surveillance}
\description{
A pipeable workflow around the 'TreeScan' command-line software:
\code{\link[=ts_visits]{ts_visits()}} -> \code{\link[=ts_counts]{ts_counts()}} -> \code{\link[=ts_run]{ts_run()}} -> \code{\link[=ts_results]{ts_results()}}.
}
\author{
\strong{Maintainer}: George G. Vega Yon \email{g.vegayon@gmail.com}

Authors:
\itemize{
  \item George G. Vega Yon \email{g.vegayon@gmail.com}
}

}
\keyword{internal}
