% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{ts_binary}
\alias{ts_binary}
\title{Locate the TreeScan command-line binary}
\usage{
ts_binary(path = getOption("treescanr.binary", Sys.getenv("TREESCAN_BIN")))
}
\arguments{
\item{path}{Path to the binary (e.g., \code{treescan64} or \code{treescan64.exe}).
Defaults to the \code{treescanr.binary} option, then the \code{TREESCAN_BIN}
environment variable, then \code{treescan64} on the \code{PATH}.}
}
\value{
The normalized path to the binary.
}
\description{
Checks that the TreeScan binary exists and is executable. The non-graphical
(command-line) version of TreeScan can be downloaded from
\url{https://www.treescan.org/download_treescan.html}.
}
\examples{
\dontrun{
options(treescanr.binary = "~/TreeScan/treescan64")
ts_binary()
}
}
