% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm.R
\name{ts_prm}
\alias{ts_prm}
\alias{ts_prm_read}
\alias{ts_prm_template}
\alias{ts_prm_set}
\alias{ts_prm_get}
\alias{ts_prm_write}
\title{TreeScan parameter files}
\usage{
ts_prm_read(path)

ts_prm_template()

ts_prm_set(prm, ...)

ts_prm_get(prm, key)

ts_prm_write(prm, path)
}
\arguments{
\item{path}{Path to a \code{.prm} file.}

\item{prm}{A \code{ts_prm} object.}

\item{...}{Named values to set, e.g. \code{"monte-carlo-replications" = 999}.
Dates are formatted as \code{yyyy/mm/dd}, logicals as \code{y}/\code{n}.}

\item{key}{Parameter name.}
}
\value{
\code{ts_prm_read()}, \code{ts_prm_template()}, and \code{ts_prm_set()} return a
\code{ts_prm} object. \code{ts_prm_get()} returns a character value.
\code{ts_prm_write()} returns \code{path} invisibly.
}
\description{
A parameter file (\code{.prm}) is stored as its raw lines (comments included)
with class \code{ts_prm}. Values are set by key name, without needing to know
the \verb{[Section]} they belong to.
}
\details{
\code{ts_prm_template()} returns the template bundled with the package
(TreeScan v2.4.1, tree-temporal Poisson scan, 9,999 replications).
}
\examples{
prm <- ts_prm_template() |>
  ts_prm_set("monte-carlo-replications" = 999)
ts_prm_get(prm, "monte-carlo-replications")
ts_prm_write(prm, tempfile(fileext = ".prm"))
}
